Split out RelativeCoordinateAutoSizingContainer into HitObjectCollection, beginning major rework/refactoring.
